uv/tests: update to incomplete marker test
I believe this is actually a correct change and reflects a fix for an
as-yet-unknown bug. Specifically, here's the corresponding packse
scenario:

```
[root]
requires = [
  "a==1; python_version < '3.10'",
  "a==2; python_version >= '3.11'",
  "b",
]

[packages.a.versions."1.0.0"]
[packages.a.versions."2.0.0"]

[packages.b.versions."1.0.0"]
requires = [
  "c; python_version == '3.10'",
]

[packages.c.versions."1.0.0"]
```

Previously, we would install `c` for *any* Python less than 3.11, but
the marker in the scenario is only for 3.10.

I don't yet have an intuition for why we were buggy before and why this
changes fixes the bug.